# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant food ordering API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: menu
    description: Menu and category operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all food items with optional filters
      parameters:
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: food_type
          in: query
          schema:
            type: ["string", "null"]
            enum: [veg, nonveg, null]
        - name: min_price
          in: query
          schema:
            type: ["number", "null"]
        - name: max_price
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new food item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "403":
          description: Forbidden

  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single food item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "404":
          description: Not found
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a food item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a food item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: Get current user's orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized

  /orders/all:
    get:
      operationId: listAllOrders
      tags: [orders]
      summary: List all orders (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (superadmin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden

  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden

  /admin/dashboard:
    get:
      operationId: getDashboardStats
      tags: [admin]
      summary: Get dashboard statistics (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "403":
          description: Forbidden

  /menu/featured:
    get:
      operationId: getFeaturedItems
      tags: [menu]
      summary: Get featured/top-rated food items for homepage
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, admin, user]
        phone:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        createdAt:
          type: string

    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [superadmin, admin, user]

    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: ["string", "null"]

    FoodItem:
      type: object
      required: [id, name, price, food_type, is_available, rating, category_id]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        food_type:
          type: string
          enum: [veg, nonveg]
        is_available:
          type: boolean
        rating:
          type: number
        category_id:
          type: integer
        category:
          $ref: "#/components/schemas/Category"

    FoodItemInput:
      type: object
      required: [name, price, food_type, category_id]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image:
          type: string
        food_type:
          type: string
          enum: [veg, nonveg]
        is_available:
          type: boolean
        category_id:
          type: integer

    FoodItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image:
          type: string
        food_type:
          type: string
          enum: [veg, nonveg]
        is_available:
          type: boolean
        category_id:
          type: integer

    OrderItem:
      type: object
      required: [food_item_id, quantity, price]
      properties:
        food_item_id:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        food_item:
          $ref: "#/components/schemas/FoodItem"

    OrderInput:
      type: object
      required: [items, address]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        address:
          type: string

    OrderItemInput:
      type: object
      required: [food_item_id, quantity]
      properties:
        food_item_id:
          type: integer
        quantity:
          type: integer

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, delivered, cancelled]

    Order:
      type: object
      required: [id, status, total_price, address, created_at]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, preparing, delivered, cancelled]
        total_price:
          type: number
        address:
          type: string
        created_at:
          type: string
        user:
          $ref: "#/components/schemas/User"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    DashboardStats:
      type: object
      required: [total_orders, total_revenue, total_users, total_menu_items, orders_by_status, recent_orders]
      properties:
        total_orders:
          type: integer
        total_revenue:
          type: number
        total_users:
          type: integer
        total_menu_items:
          type: integer
        orders_by_status:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusCount"
        recent_orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    OrderStatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer