# lib/api-zod/src/generated/types/foodItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant food ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodItemInputFoodType } from './foodItemInputFoodType';

export interface FoodItemInput {
  name: string;
  description?: string;
  price: number;
  image?: string;
  food_type: FoodItemInputFoodType;
  is_available?: boolean;
  category_id: number;
}