# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant food ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';
import type { User } from './user';

export interface Order {
  id: number;
  status: OrderStatus;
  total_price: number;
  address: string;
  created_at: string;
  user?: User;
  items?: OrderItem[];
}