# lib/api-zod/src/generated/types/userRoleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant food ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRoleUpdateRole } from './userRoleUpdateRole';

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}